virheilmoitukset päänäkymään

+++ harjoitustyo/src/front_view.py
@@ -1,4 +1,4 @@
-from tkinter import ttk, constants
+from tkinter import ttk, constants, StringVar
 
 class DataView:
     '''
@@ -21,6 +21,7 @@ class DataView:
         self._root = root
         self._frame = None
         self._service = service
+        self._error_variable = None
 
         self._initialize()
 
@@ -44,13 +45,25 @@ class DataView:
         estimation = self._service.get_estimation()
         data = self._service.print_data()
 
+        self._error_variable = StringVar(self._frame)
+
         heading_label = ttk.Label(master=self._frame, text="Varallisuusarvio")
         estimation_label = ttk.Label(master=self._frame, text=estimation)
         data_label = ttk.Label(master=self._frame, text=data)
+        self._error_label = ttk.Label(master=self._frame, textvariable=self._error_variable, foreground="red")
 
         heading_label.grid(row=0, column=0, sticky=constants.W)
         estimation_label.grid(row=1, column=0, sticky=constants.W)
         data_label.grid(row=2, column=0, sticky=constants.W)
+
+        self._hide_error()
+
+    def _show_error(self, message):
+        self._error_variable.set(message)
+        self._error_label.grid(row=3, column=2)
+
+    def _hide_error(self):
+        self._error_label.grid_remove()
 
 class FrontView:
     '''
@@ -60,6 +73,8 @@ class FrontView:
         root: Sovelluksen ikkuna.
         service: Sovelluslogiikasta huolehtiva luokka.
         show_login_view: Metodi, joka näyttää kirjautumisnäymän.
+        error_variable: Virheviesti
+        error_label: Virhe
     '''
 
     def __init__(self, root, service, show_login_view):
@@ -70,6 +85,8 @@ class FrontView:
             root: Sovelluksen ikkuna.
             service: Sovelluslogiikasta huolehtiva luokka.
             show_login_view: Metodi, joka näyttää kirjautumisnäymän.
+            error_variable: Virheviesti
+            error_label: Virhe
         '''
 
         self._root = root
@@ -162,35 +179,45 @@ class FrontView:
 
         self._data_frame.grid(row=0, column=0)
 
-        name1_label.grid(row=3, column=0, sticky=constants.E)
-        amount1_label.grid(row=4, column=0, sticky=constants.E)
+        name1_label.grid(row=4, column=0, sticky=constants.E)
+        amount1_label.grid(row=5, column=0, sticky=constants.E)
 
-        self._add_income_name_entry.grid(row=3, column=1)
-        self._add_income_amount_entry.grid(row=4, column=1)
-        add_income_button.grid(row=5, column=1)
+        self._add_income_name_entry.grid(row=4, column=1)
+        self._add_income_amount_entry.grid(row=5, column=1)
+        add_income_button.grid(row=6, column=1)
 
-        self._add_expense_name_entry.grid(row=3, column=2)
-        self._add_expense_amount_entry.grid(row=4, column=2)
-        add_expense_button.grid(row=5, column=2)
+        self._add_expense_name_entry.grid(row=4, column=2)
+        self._add_expense_amount_entry.grid(row=5, column=2)
+        add_expense_button.grid(row=6, column=2)
 
-        name2_label.grid(row=6, column=0, sticky=constants.E)
+        name2_label.grid(row=7, column=0, sticky=constants.E)
 
-        self._delete_income_name_entry.grid(row=6, column=1)
-        delete_income_button.grid(row=7, column=1)
+        self._delete_income_name_entry.grid(row=7, column=1)
+        delete_income_button.grid(row=8, column=1)
 
-        self._delete_expense_name_entry.grid(row=6, column=2)
-        delete_expense_button.grid(row=7, column=2)
+        self._delete_expense_name_entry.grid(row=7, column=2)
+        delete_expense_button.grid(row=8, column=2)
 
-        amount2_label.grid(row=8, column=0, sticky=constants.E)
+        amount2_label.grid(row=9, column=0, sticky=constants.E)
 
-        self._update_wealth_entry.grid(row=8, column=1)
-        update_wealth_button.grid(row=9, column=1)
+        self._update_wealth_entry.grid(row=9, column=1)
+        update_wealth_button.grid(row=10, column=1)
 
-        delete_all_data_button.grid(row=9, column=2)
+        delete_all_data_button.grid(row=10, column=2)
 
     def _handle_add_income(self):
         add_income_name = self._add_income_name_entry.get()
-        add_income_amount = int(self._add_income_amount_entry.get())
+        add_income_amount = self._add_income_amount_entry.get()
+
+        if not add_income_amount or not add_income_name:
+            self._data_view._show_error('Täytä molemmat kentät')
+            return
+
+        try:
+            add_income_amount = int(add_income_amount)
+        except ValueError:
+            self._data_view._show_error('Anna summa kokonaislukuna')
+            return
 
         self._service.add_income(add_income_name, add_income_amount)
 
@@ -198,7 +225,17 @@ class FrontView:
 
     def _handle_add_expense(self):
         add_expense_name = self._add_expense_name_entry.get()
-        add_expense_amount = int(self._add_expense_amount_entry.get())
+        add_expense_amount = self._add_expense_amount_entry.get()
+
+        if not add_expense_amount or not add_expense_name:
+            self._data_view._show_error('Täytä molemmat kentät')
+            return
+
+        try:
+            add_expense_amount = int(add_expense_amount)
+        except ValueError:
+            self._data_view._show_error('Anna summa kokonaislukuna')
+            return
 
         self._service.add_expense(add_expense_name, add_expense_amount)
 
@@ -207,6 +244,10 @@ class FrontView:
     def _handle_delete_income(self):
         delete_income_name = self._delete_income_name_entry.get()
 
+        if not self._service.search_income(delete_income_name):
+            self._data_view._show_error('Tuloa ei löytynyt')
+            return
+
         self._service.delete_income(delete_income_name)
 
         self._initialize_data()
@@ -214,12 +255,22 @@ class FrontView:
     def _handle_delete_expense(self):
         delete_expense_name = self._delete_expense_name_entry.get()
 
+        if not self._service.search_expense(delete_expense_name):
+            self._data_view._show_error('Menoa ei löytynyt')
+            return
+
         self._service.delete_expense(delete_expense_name)
 
         self._initialize_data()
 
     def _handle_update_wealth(self):
-        wealth = int(self._update_wealth_entry.get())
+        wealth = self._update_wealth_entry.get()
+
+        try:
+            wealth = int(wealth)
+        except ValueError:
+            self._data_view._show_error('Anna summa kokonaislukuna')
+            return
 
         self._service.update_wealth(wealth)
